Немного перетасовал классы по файлам. Теперь классы стратегий подключаются раньше, чем классы использующие их. Конечные классы элементов перемещены к абстрактным классам, а базовые классы интегрирования ближе к конечным классам интегрирования. Экспериментальный функционал для кубических серендиповых элементов перенесён в папку experimental до лучших времён.
